requirements.txt
django==3.0.5
pillow
django-allauth
tk
channels
pandas
openpyxl
whitenoise